Add AffineLoopInterchange pass
Adds a loop interchange pass to the Affine dialect.
The pass works only for loopnests having rectangular iteration space and no Affine-If statements.
In all other cases, it detects the absence of the above conditions and simply bails out leaving that particular loopnest unchanged.
The pass works for both - imperfectly nested loop nests as well as perfectly nested ones.